# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "سیستم مدیریت نوبت‌دهی کلینیک دندانپزشکی",
    "description": "API کامل و حرفه‌ای برای مدیریت نوبت‌دهی دو کلینیک دندانپزشکی\n\n⚠️ نکته مهم CSRF Protection:\n1. ابتدا از /auth/csrf-token یک token دریافت کنید\n2. در header درخواست‌های POST/PATCH/DELETE، X-CSRF-Token را با مقدار token ارسال کنید",
    "version": "1.4.0",
    "contact": {
      "name": "Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Alternative port"
    }
  ],
...
  "paths": {
...
    "/services": {
      "get": {
        "tags": ["Services"],
        "summary": "لیست خدمات",
        "operationId": "getServices",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "جستجو در عنوان، توضیحات و نکات"
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "description": "فیلتر بر اساس ID دسته‌بندی"
          },
          {
            "name": "categorySlug",
            "in": "query",
            "schema": { "type": "string" },
            "description": "فیلتر بر اساس slug دسته‌بندی"
          }
        ],
        "responses": {
          "200": {
            "description": "لیست خدمات",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Service" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Services"],
        "summary": "ایجاد خدمت (مدیر/منشی)",
        "operationId": "createService",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["title", "description"],
                "properties": {
                  "title": { "type": "string", "example": "جراحی ایمپلنت" },
                  "description": { "type": "string", "example": "توضیحات خدمت" },
                  "beforeTreatmentTips": { "type": "string", "example": "نکات قبل از درمان" },
                  "afterTreatmentTips": { "type": "string", "example": "نکات بعد از درمان" },
                  "price": { "type": "integer", "example": 5000000 },
                  "durationMinutes": { "type": "integer", "example": 120 },
                  "categoryIds": { 
                    "type": "array", 
                    "items": { "type": "string", "format": "uuid" },
                    "description": "آرایه ID دسته‌بندی‌ها (اختیاری) - می‌تواند یک UUID یا آرایه باشد"
                  },
                  "coverImage": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "خدمت ایجاد شد" }
        }
      }
    },
    "/services/{identifier}": {
      "get": {
        "tags": ["Services"],
        "summary": "دریافت خدمت (با slug یا ID)",
        "operationId": "getService",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "خدمت" }
        }
      }
    },
    "/services/{id}": {
      "patch": {
        "tags": ["Services"],
        "summary": "به‌روزرسانی خدمت (مدیر/منشی)",
        "operationId": "updateService",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "beforeTreatmentTips": { "type": "string" },
                  "afterTreatmentTips": { "type": "string" },
                  "price": { "type": "integer" },
                  "durationMinutes": { "type": "integer" },
                  "categoryIds": { 
                    "type": "array", 
                    "items": { "type": "string", "format": "uuid" },
                    "description": "آرایه ID دسته‌بندی‌ها (اختیاری) - می‌تواند یک UUID، آرایه یا [] برای حذف همه باشد"
                  },
                  "coverImage": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "خدمت به‌روزرسانی شد" }
        }
      },
      "delete": {
        "tags": ["Services"],
        "summary": "حذف خدمت (مدیر/منشی)",
        "operationId": "deleteService",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "خدمت حذف شد" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "beforeTreatmentTips": { "type": "string" },
          "afterTreatmentTips": { "type": "string" },
          "price": { "type": "integer" },
          "durationMinutes": { "type": "integer" },
          "coverImage": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "name": { "type": "string" },
                "slug": { "type": "string" }
              }
            },
            "description": "لیست دسته‌بندی‌های خدمت"
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "شناسه یکتای رکورد"
      },
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        },
        "description": "شماره صفحه"
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        },
        "description": "تعداد رکوردها در هر صفحه"
      }
    },
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dental.sid",
        "description": "Session-based authentication using cookies"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token برای حفاظت در برابر حملات CSRF. ابتدا /auth/csrf-token را صدا کنید و token دریافتی را در این header ارسال کنید"
      }
    }
  }
}